refactor(application): move application request to application module
