+++ backend/src/modules/application/application.module.ts
@@ -8,19 +8,23 @@ import { SharedModule } from 'src/shared/shared.module';
 import { OngApplication } from './entities/ong-application.entity';
 import { OngApplicationRepository } from './repositories/ong-application.repository';
 import { OngApplicationService } from './services/ong-application.service';
-import { OrganizationApplicationController } from './controllers/organization-application.controller';
+import { ApplicationRequest } from './entities/application-request.entity';
+import { ApplicationRequestRepository } from './repositories/application-request.repository';
+import { ApplicationRequestService } from './services/application-request.service';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Application, OngApplication]),
+    TypeOrmModule.forFeature([Application, OngApplication, ApplicationRequest]),
     SharedModule,
   ],
-  controllers: [ApplicationController, OrganizationApplicationController],
+  controllers: [ApplicationController],
   providers: [
     ApplicationService,
     ApplicationRepository,
     OngApplicationRepository,
     OngApplicationService,
+    ApplicationRequestRepository,
+    ApplicationRequestService,
   ],
   exports: [OngApplicationService, ApplicationService],
 })
