+++ backend/src/modules/application/constants/application-filters.config.ts
@@ -19,3 +19,31 @@ export const APPLICATION_FILTERS_CONFIG = {
   defaultOrderDirection: OrderDirection.ASC,
   relations: {},
 };
+
+export const APPLICATION_REQUEST_FILTERS_CONFIG = {
+  selectColumns: {
+    id: true,
+    application: {
+      id: true,
+      name: true,
+    },
+    organization: {
+      id: true,
+      organizationGeneral: {
+        name: true,
+      },
+    },
+    status: true,
+    createdOn: true,
+  },
+  searchableColumns: [],
+  defaultSortBy: 'createdOn',
+  defaultOrderDirection: OrderDirection.ASC,
+  relations: {
+    organization: {
+      organizationGeneral: true,
+    },
+    application: true,
+  },
+  rangeColumn: 'createdOn',
+};
