+++ backend/src/modules/requests/constants/request-filters.config.ts
@@ -28,31 +28,3 @@ export const REQUEST_FILTER_CONFIG = {
   },
   rangeColumn: 'createdOn',
 };
-
-export const REQUEST_APP_ACCESS_FILTER_CONFIG = {
-  selectColumns: {
-    id: true,
-    application: {
-      id: true,
-      name: true,
-    },
-    organization: {
-      id: true,
-      organizationGeneral: {
-        name: true,
-      },
-    },
-    status: true,
-    createdOn: true,
-  },
-  searchableColumns: [],
-  defaultSortBy: 'createdOn',
-  defaultOrderDirection: OrderDirection.ASC,
-  relations: {
-    organization: {
-      organizationGeneral: true,
-    },
-    application: true,
-  },
-  rangeColumn: 'createdOn',
-};
